Feature/implement ad and purchase logic

+++ src/main/java/br/com/mercadolivre/projetointegrador/marketplace/services/AdService.java
@@ -1,21 +1,60 @@
 package br.com.mercadolivre.projetointegrador.marketplace.services;
 
+import br.com.mercadolivre.projetointegrador.marketplace.dtos.CreateOrUpdateAdDTO;
+import br.com.mercadolivre.projetointegrador.marketplace.exceptions.UnauthorizedException;
 import br.com.mercadolivre.projetointegrador.marketplace.model.Ad;
+import br.com.mercadolivre.projetointegrador.marketplace.model.AdBatch;
+import br.com.mercadolivre.projetointegrador.marketplace.repository.AdBatchesRepository;
 import br.com.mercadolivre.projetointegrador.marketplace.repository.AdRepository;
 import br.com.mercadolivre.projetointegrador.warehouse.exception.db.NotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
 @AllArgsConstructor
+@Transactional
 public class AdService {
 
   AdRepository adRepository;
+  AdBatchesRepository adBatchesRepository;
 
-  public void createAd(Ad ad) {
-    this.adRepository.save(ad);
+  @Transactional(
+      rollbackFor = Exception.class,
+      propagation = Propagation.REQUIRED,
+      isolation = Isolation.SERIALIZABLE)
+  public Ad createAd(Long sellerId, CreateOrUpdateAdDTO createAdDTO) {
+    Ad ad = new Ad();
+    ad.setSellerId(sellerId);
+    ad.setName(createAdDTO.getName());
+    ad.setQuantity(createAdDTO.getQuantity());
+    ad.setPrice(createAdDTO.getPrice());
+    ad.setDiscount(createAdDTO.getDiscount());
+    ad.setCategory(createAdDTO.getCategory());
+
+    List<Long> batchesId = createAdDTO.getBatchesId();
+    adRepository.save(ad);
+
+    for (Long id : batchesId) {
+      AdBatch adBatch = new AdBatch();
+      adBatch.setBatchId(id);
+      adBatch.setAd(ad);
+      adBatchesRepository.save(adBatch);
+    }
+
+    return ad;
+  }
+
+  public List<Ad> listAds() {
+    return this.adRepository.findAll();
+  }
+
+  public List<Ad> listAds(String name) {
+    return this.adRepository.findAdsByLikeName(name);
   }
 
   public List<Ad> listAdsByCustomerId(Long id) {
@@ -30,13 +69,11 @@ public class AdService {
     return ad;
   }
 
-  public void updateAd(Ad updatedAd) {
-    findAdById(updatedAd.getId());
-    adRepository.save(updatedAd);
-  }
-
-  public void deleteAd(Long id) throws NotFoundException {
-    Ad ad = findAdById(id);
-    adRepository.delete(ad);
+  public void deleteAd(Long customerId, Long adId) throws UnauthorizedException {
+    Ad ad = findAdById(adId);
+    if (ad.getSellerId().equals(customerId)) {
+      adRepository.delete(ad);
+    }
+    throw new UnauthorizedException("Não é permitido excluir o anúncio de outro usuário.");
   }
 }
